Lanugage  Writer Package Refactor
1) Decomposed the existing language writer interface into smaller
components
2) Updated C and C++ lanugage writer implementations to conform to the
new interface structure
3) Updated project configurations to control the new lanugage writers
