+++ pal/config.py
@@ -115,7 +115,7 @@ config.add_configuration(c)
 
 c = Configuration("language", "c++11")
 c.description = "The target programming language"
-c.options = ["c++11", "yaml", "none"]
+c.options = ["c", "c++11", "yaml", "none"]
 config.add_configuration(c)
 
 c = Configuration("access_mechanism", "gas_att")
@@ -124,8 +124,8 @@ c.options = ["gas_att", "gas_intel", "gas_aarch64", "gas_aarch32", "test",
              "yaml", "none"]
 config.add_configuration(c)
 
-c = Configuration("print_style", "printf_utf8")
-c.description = "The target function printing style"
+c = Configuration("print_mechanism", "printf_utf8")
+c.description = "The target mechanism for printing messages"
 c.options = ["printf_utf8", "none"]
 config.add_configuration(c)
 
@@ -136,7 +136,7 @@ config.add_configuration(c)
 
 c = Configuration("generator", "c++_header")
 c.description = "The target generator"
-c.options = ["c++_header", "yaml"]
+c.options = ["c_header", "c++_header", "yaml"]
 config.add_configuration(c)
 
 # -----------------------------------------------------------------------------
@@ -204,7 +204,7 @@ config.add_configuration(c)
 # C options
 # -----------------------------------------------------------------------------
 
-c = Configuration("c_prefix", "")
+c = Configuration("c_prefix", "pal")
 c.description = "Prefix to place in front of generated C functions"
 config.add_configuration(c)
