+++ pal/writer/access_mechanism/none.py
@@ -4,9 +4,6 @@ from pal.writer.access_mechanism.access_mechanism \
 
 class NoneAccessMechanismWriter(AccessMechanismWriter):
 
-    def declare_access_mechanism_dependencies(self, outfile, register):
-        pass
-
     def call_readable_access_mechanism(self, outfile, register,
                                        access_mechanism, var):
         pass
