+++ pal/writer/language/none.py
@@ -4,19 +4,13 @@ from pal.writer.language.language import LanguageWriter
 
 class NoneLanguageWriter(LanguageWriter):
 
-    def declare_register_constants(self, outfile, register):
-        pass
-
     def declare_comment(self, outfile, comment, wrap=79):
         pass
 
-    def declare_register_get(self, outfile, register):
+    def declare_register_dependencies(self, outfile, register):
         pass
 
-    def declare_register_set(self, outfile, register):
-        pass
-
-    def declare_field_constants(self, outfile, register, field):
+    def declare_register_accessors(self, outfile, register):
         pass
 
     def declare_field_accessors(self, outfile, register, field):
@@ -25,5 +19,6 @@ class NoneLanguageWriter(LanguageWriter):
     def call_register_get(self, outfile, register, destination, index="index"):
         pass
 
-    def call_field_get(self, outfile, register_value, field, destination):
+    def call_field_get(self, outfile, register, field, destination,
+                       register_value):
         pass
